requirements.txt
cryptography>=41.0.7
python-dotenv>=1.0.0
pynput>=1.7.6
psutil>=5.9.6
requests>=2.31.0
colorama>=0.4.6
scikit-learn>=1.3.2
pandas>=2.1.4
numpy>=1.26.2
aiohttp>=3.9.1
websockets>=12.0
pytest>=7.4.3
pytest-asyncio>=0.23.2
pytest-cov>=4.1.0
pyyaml==6.0.1
sphinx>=7.1.2
sphinx-rtd-theme>=1.3.0
prometheus-client>=0.19.0
PyJWT>=2.8.0
bcrypt>=4.1.1
redis>=5.0.1
asyncpg>=0.29.0
pyautogen>=0.2.0
azure-identity==1.15.0
azure-storage-blob==12.19.0
azure-cosmos==4.5.1
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-keyvault-secrets==4.7.0
azure-core==1.29.7
asyncio==3.4.3
fastapi>=0.104.1
uvicorn>=0.24.0
spacy>=3.7.2
textblob>=0.17.1
transformers>=4.35.2
torch>=2.1.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
